# generated-resources/src/main/resources/openapi.yaml
openapi: 3.0.0
info:
  title: Combined SimilarProducts API
  version: "1.0"
servers:
- url: http://localhost:5000
- url: http://localhost:3001
paths:
  /product/{productId}/similar:
    get:
      operationId: getSimilarProductDetails
      parameters:
      - explode: false
        in: path
        name: productId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilarProductDetails'
          description: OK
        "404":
          description: Product Not found
      summary: Similar products
      x-accepts: application/json
  /product/{productId}/similarids:
    get:
      operationId: getSimilarProductsIds
      parameters:
      - explode: false
        in: path
        name: productId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SimilarProductIds'
          description: OK
      summary: Gets the ids of similar products
      x-accepts: application/json
  /product/{productId}:
    get:
      operationId: getProductId
      parameters:
      - explode: false
        in: path
        name: productId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDetailBasic'
          description: OK
        "404":
          description: Product Not found
      summary: Gets a product detail
      x-accepts: application/json
components:
  schemas:
    SimilarProductDetails:
      description: List of similar products with full details
      items:
        $ref: '#/components/schemas/ProductFullDetail'
      type: array
    ProductFullDetail:
      description: Product detail
      example:
        price: 0.8008281904610115
        name: name
        id: id
        availability: true
      properties:
        id:
          minLength: 1
          type: string
        name:
          minLength: 1
          type: string
        price:
          type: number
        availability:
          type: boolean
      required:
      - availability
      - id
      - name
      - price
      type: object
    SimilarProductIds:
      description: List of similar product IDs
      items:
        type: string
      type: array
    ProductDetailBasic:
      description: Product detail
      example:
        price: 0.8008281904610115
        name: name
        id: id
        availability: true
      properties:
        id:
          minLength: 1
          type: string
        name:
          minLength: 1
          type: string
        price:
          type: number
        availability:
          type: boolean
      required:
      - availability
      - id
      - name
      - price
      type: object